Change consumer attributes field to textarea
SCRD-2241 The consumer attributes for a device group in the Disk Models tab
of the Manage Cloud Setting modal was not able to accommodate different
forms of consumer attributes because it was a text input field. Changed it
to a textarea input field and added logic to handle the conversion between
yaml and javascript object.