InterfaceModels: Enable save only when something has changed
Modify the logic that controls the Save button on the Inteface Models tab of
the cloud settings dialog so that it is enabled only when the entries
are valid and when something has changed.  Previously the Save button
was enabled solely based on whether the data was valid, which led to it
being available even when nothing had changed, which may be slightly
confusing to the user.