Add Edit Server modal for servers added manually (bsc#1080193)
Bug 1080193 - Added the Edit Server modal for servers added manually.
The work includes refactoring out the existing code that handles servers
added manually to be it's own module and reusing it for the edit modal.

Also fixed the View Details modal so that it's look and feel is consistent
with the Add and Edit modals.

The changes in CloudModalSummary and CsvImporter are fixes for lints complaints.