Handle custom model names in selection page (bsc#1082586)
If the user has selected a model and progressed through the installer
and changed then name on the page that permits model editing and then
returned back to the model selection page, display (and automatically
select) a new "custom" box that reflects this change.  This permits the
user to go forward again without forcing a selection of an existing
template and overwriting their customized changes to the model.