SCRD-2374 Support cancelling playbooks (#89)
Simplify findNextPlaybook to expect just a single playbook (since it
really does not expect a full list of completed playbooks), and change
sumPlaybookProcessStatus to a simpler function that gets playbooks with
the given status.

Simplify logic:
Adjust the redundant amount of if/else nesting in processPlaybooks
Move the common logic for processing completed playbooks out of the
if/then/else since all branches had that as their first step. Combine the logic for two of the branches into one.

Display the error message only after cancelling, and avoid displaying
both an error message and the log.