SCRD-3458 Create menu system
Create navigational menu system in the style of the EOS Design System
[1].  Uses react-router to create a route to each of the main menu items
and sub-items.  This version supports only a fixed, static menu
structure in order to centralize the logic for routing and menu
processing.  If we later need to perform dynamic routing (and dynamic
menus), then this can be added but may complicate the logic a bit.

[1] git@gitlab.com:SUSE-UIUX/eos.git