Include Lato font files with install UI (bsc#1100909)
Currently the Install UI is trying to download the
Lato font from googles font repository online,
but the Install UI is supposed to be runnable
without an internet connection. The result is that
if you're disconnected from the internet the UI
loads *very* slowly (waits for timeout on the font
load).

This change adds the Lato font ttf file to the
Install UI so it can be loaded locally, and
removes the remote reference URL from the
css file.