Add Services Per Role page (#119)
* Add Services Per Role page

SCRD-3488 Added 'Service Per Role' which is the page under a new tab called
'Roles' off of the Services menu. The page contains three columns: Roles, Servers,
and Services. The Servers column shows host IDs instead of host names at the
moment. The plan is to show host names when we can get them from the output of
the config processor run which will come later on.