Add Service Information page
SCRD-3484 Added 'Service Information' page which is the page under
the 'Information' tab off of the Services menu. The page contains a
table with four columns: Name, Description, Endpoints, and Regions.
There's supposed to be the fifth column which is the 'Status' column,
But we need the status data from Monasca which is not available at
the moment. Once the Monasca data is available, we will need to add
the Status column to the table.