Rebased with latest master and added a summary tab
Implemented framework for replacing servers

1. Added servers landing page to show all the servers. Made it
   interact with the menu.
2. Added context menu for the page and updated context menu
   to be functioning
3. Updated Install wizard to deal with update page and refresh
   update page whilt it is still in progress
4. Added replace server modal
5. Moved server details component to the components dir
6. Moved CollapsibleTable to the components dir
7. Added 3 placeholder pages to make sure the flow works

Note:
There are lots of TODOs need to be resolved.
Select from available servers need further work.
There is no restriction for multiple users running the update,
need to think about the design to restrict the funcitonalities
when updating is in progress.
The progress.json need to be cleaned before start day2 UI

To test:
1. Test without available servers
   you can just enter any ilo inputs

2. Test with available servers
   need to go to day0 to get some manual servers
   or discovered servers, they will  show up in the dropdown
   list in replace detail modal