Pass known servers to ReplaceServerDetails
Before calling ReplaceServerDetails, create an array which contains
the set of all servers that are "known", which are those that have
been discovered or manually entered.  This array is sent as a property,
which can then be relied upon to be an defined as a (possibly empty)
array.  This simplifies code by removing logic throughout
ReplaceServerDetails which repeatedly tried to handle two possibly
undefined arrays and concatenate them together.  This also removed
the need to use a couple of custom functions in ModelUtils that
do little more than repeat those same checks and concatenation.