SCRD-5072 Improve replace controller dialog (#179)
SCRD-5072 Improve replace controller dialog

* Combine rendering of replacement OS user and password

Refactor the code to combine the rendering of the OS username and
OS password on the replace server dialog into a single function.  This
simplifies the logic to control whether these fields are shown.

* Remove class variable, refactor getAvailableServerIds

Remove the availableServerIds class variable and calculate it as needed.
Remove getAvailableServerIds, which was only used in one place, from
ModelUtils and incorporate a simplified version of its logic inline.

* Pass known servers to ReplaceServerDetails

Before calling ReplaceServerDetails, create an array which contains
the set of all servers that are "known", which are those that have
been discovered or manually entered.  This array is sent as a property,
which can then be relied upon to be an defined as a (possibly empty)
array.  This simplifies code by removing logic throughout
ReplaceServerDetails which repeatedly tried to handle two possibly
undefined arrays and concatenate them together.  This also removed
the need to use a couple of custom functions in ModelUtils that
do little more than repeat those same checks and concatenation.

* Simplify logic for rendering server details table

* Remove remaining references to this.data

The 'data' member of the object had no valid reason to exist, so
removing it.

* Fix state handling