SCRD-3497 Support running non-playbooks with PlaybookProgress
Add support for supplying an arbitrary action instead of only launching
ansible playbooks. The action is a function that returns a promise,
which enables it to be used for making API calls to the back end as well
as any other call.
Replaced logic that used isUpdateMode to indicate that the playbook
array was to contain objects (instead of just names) with logic that
inspected the array and made the decision based on the contents.
Added documentation for the props passed in to PlaybookProcess.
Added an example function in PrepareReplace.