SCRD-3497 Add PrepareController
Add a new page for the prepare server workflow that is specifically for
preparing control nodes for replacement, up to the point where the OS
installation would begin.  Moved the model commit into an action so that
it will not be repeated when the page is realoaded.
Moved PrepareReplace to PrepareCompute, since the workflow for replacing
a compute node is different.