SCRD-3497 Reconfigure ceilometer
When a controller node is replaced, ceilometer may need to be
reconfigured based on the configuration.  Per section 12.3.5.5 of the
operations guide, this is only necessary when the controller node is a
dedicated MML node, but since the playbooks are idempotent, there is no
real downside to running them everywhere.