SCRD-5562 Stop tracking inputValue in internal state
Instead of tracking the inputValue in both the internal state as well as
the props, rely on the parent to update the props whenever the user
changes the value, which is done via the inputAction callback.  This
matches the more standard reactJS model and fixes spurious problems
seen when updating the control programatically.