SCRD-3111 Use smarter caching of internal model
Loading the model that is output from the config processor, a.k.a., the
internal model, was potentially quite slow (up to about 10 seconds), so
the getInternalModel function was originally created in order to avoid
even re-querying the back-end, making the UI much faster.  But that
solution incorrectly served up old results after the config processor
was run again.  So some pages were bypassing getInternalModel and
performing the fetch directly and explicitly adding 'cache-control:
no-cache' in order to forcably re-load fresh results, and other pages
were not.

Re-factor the code to always use the no-cache header when reading the
internal model.  Note that the no-cache option has a misleading name,
since it actually compares the cached response with the ETag from the
server and only re-loads if necessary.