SCRD-3497 Avoid race condition with model save and commit
The step to commit the model did not first wait for the model to be
saved, and if the commit happened first, then uncommitted changes would
result, causing the replace controller flow to terminate.  Modify the
updateGlobalState function to return a promise that is resolved when
the save is complete, so that the following functions can be
synchronized properly.