(Re-)join room at connect time
By default the socket IO client library will detect interruptions in
communications with the server and automatically retry continually.
But when connectivity is re-estiblished, the client is not automitically
re-added to room in which it was listening for messages.  This change
moves the code to join the room into the callback function that is
triggered whenever the connection (or re-connection) succeeds.