SCRD-4240 Add passphrase dialog to login page
At the login prompt add a modal dialog to prompt for the SSH passphrase
if one is needed in order to load the private SSH key.  Improve the
text to explain how it is used and how to deal with restarts while
logged in.  Remove the ability to close the dialog without entering
a valid passphrase. Remove the ssh password dialog from the ServiceInfo
page.