Add ipv6 support (SOC-9677) (#357)
Incorporate third-party library, ipaddr.js, to implement various
functions that validate and process IP addresses and CIDRs, which are
more advanced than simple regex matches. Surround literal ipv6 addresses
with brackets with building URLs.

Replace the only instance of subnet+netmask fields with a CIDR field.
This was the "Manage subnet and netmask" button, now renamed to "Manage
network range", which now prompts for a single CIDR field.  This network
range is saved into servers.yml in the input model.  The network range
is still required to be separate ipv4 subnet/netmask fields in
servers.yml, so this two-way conversion is handled by the UI.  In the
future when a single CIDR field is supported in servers.yml, then the
conversion code can be removed.